refactor/#269 ServiceType과 ViewModel 병합
